Record Your Plot History

Record Your Plot History

plothistory(
  phdir = phist_tmp_dir(),
  sleep = 1,
  hgd = TRUE,
  port = 5900,
  token = FALSE,
  ...
)

Arguments

phdir

character(1) defining the plothistory directory. Default is to phist_tmp_dir().

sleep

numeric(1) > 0 indicating the clock cycle in seconds. at which phdir should be updated. Default is 1 second. Passed to Sys.sleep().

hgd

logical(1) defining whether a web graphics device should be started with port and token. Default is TRUE. In case the backgroup script stops, the plotting history can be restarted by setting hgd = FALSE.

port

numeric(1) defining the server port, passed to httpgd::hgd(). If this is set to ‘0’, an open port will be assigned. Default is 5900.

token

Optional security token, passed to httpgd::hgd(). Default is FALSE to deactivate the token.

...

Additional parameters passed to httpgd::hgd().

Value

The value of the error code (0 for success) of the background script. See system2() for details.